Route geolocation requests through internal API
