+++ .env.example
@@ -17,7 +17,7 @@ OPENAI_API_URL="https://api.openai.com/v1/chat/completions"
 NEXT_PUBLIC_JITSI_DOMAIN="meet.jit.si"
 NEXT_PUBLIC_QUOTE_API="https://api.quotable.io/random"
 OPEN_LIBRARY_API_URL="https://openlibrary.org/subjects"
-NEXT_PUBLIC_GEOLOCATION_URL="https://api.bigdatacloud.net/data/reverse-geocode-client"
+GEOLOCATION_API_URL="https://api.bigdatacloud.net/data/reverse-geocode-client"
 NEXT_PUBLIC_JITSI_DOMAIN="https://meet.jit.si"
 NEXT_PUBLIC_JOB_MATCH_API="https://api.adzuna.com/v1/api/jobs"
 AUTH_BYPASS="false"
