+++ lib/utils/location.ts
@@ -1,12 +1,7 @@
 export async function reverseGeocode(lat: number, lon: number): Promise<string> {
-  const base = process.env.NEXT_PUBLIC_GEOLOCATION_URL;
-  if (!base) throw new Error("Missing geolocation endpoint");
-  const url = `${base}?latitude=${lat}&longitude=${lon}&localityLanguage=en`;
+  const url = `/api/express/location?lat=${lat}&lon=${lon}`;
   const res = await fetch(url);
   if (!res.ok) throw new Error("Failed to fetch location");
   const data = await res.json();
-  const parts = [data.city || data.locality, data.principalSubdivision, data.countryName]
-    .filter(Boolean)
-    .join(", ");
-  return parts;
+  return data.location;
 }
